// src/enums/pokemonTypesEnum.ts
/**
 * Enum representing the original Pokémon types
 */
export enum OriginalPokemonTypesEnum {
	DARK = 'dark',
	DRAGON = 'dragon',
	ELECTRIC = 'electric',
	FIGHTING = 'fighting',
	FIRE = 'fire',
	GRASS = 'grass',
	NORMAL = 'normal',
	PSYCHIC = 'psychic',
	STEEL = 'steel',
	WATER = 'water',
	NULL = 'null',
}